Re-add PHOTON_BUILD_DEPENDENCIES in x86 CI workflow (#610)

+++ io/iouring-wrapper.cpp
@@ -188,7 +188,7 @@ public:
         ioCtx timer_ctx(true, false);
         __kernel_timespec ts;
         auto usec = timeout.timeout_us();
-        if (usec < std::numeric_limits<int64_t>::max()) {
+        if (usec < (uint64_t) std::numeric_limits<int64_t>::max()) {
             sqe->flags |= IOSQE_IO_LINK;
             ts = usec_to_timespec(usec);
             sqe = _get_sqe();
@@ -334,7 +334,7 @@ public:
 
     ssize_t wait_and_fire_events(uint64_t timeout) override {
         // Prepare own timeout
-        if (timeout > std::numeric_limits<int64_t>::max()) {
+        if (timeout > (uint64_t) std::numeric_limits<int64_t>::max()) {
             timeout = std::numeric_limits<int64_t>::max();
         }
